S3画像ファイル変更

+++ app/Http/Controllers/BookController.php
@@ -111,37 +111,61 @@ class BookController extends Controller
      */
     public function update(Request $request, $id)
     {
+      // 画像保存ディレクトリ設定
+      $save_directory = "book_images";
+      // 対象レコード取得
       $book = Bookdata::find($id);
+      // リクエストデータ受取
       $form = $request->all();
+      // フォームトークン削除
       unset($form['_token']);
 
-      // 写真削除処理
+      // 写真変更がある場合
       if (isset($form['picture'])) {
-        // 削除写真名取得
-        $deletename = $book->picture;
-        // 写真削除
-        $pathdel = storage_path() . '/app/public/book_images/' . $deletename;
-        \File::delete($pathdel);
-        // 削除のみ時
+        // 削除処理実施
+        // 開発環境で画像保存先を変更
+        if ( app()->isLocal() || app()->runningUnitTests() ) {
+          // 写真削除情報取得
+          $deletename = str_replace('/storage/'.$save_directory.'/','',$book->picture);
+          $pathdel = storage_path() . '/app/public/book_images/' . $deletename;
+          // 写真削除
+          \File::delete($pathdel);
+        } else {
+          // 本番環境削除処理
+          // S3インスタンス生成
+          $s3settings = s3settings();
+          // S3削除処理
+          $picture_delete = picture_delete($save_directory,$book,$s3settings);
+        }
+        // 削除のみか判定
         if ($form['picture'] == "no-picture") {
-          // レコード更新処理
-          $form['picture'] = null;
-        // 変更後画像がある時
+          // レコードをnull化
+          $picture_upload = null;
+        // 画像変更時
         } else {
           // 写真追加処理
-          // 画像情報取得
-          $file = $request->file('picture');
-          // 拡張子取得
-          $ext = $file->getClientOriginalExtension();
-          // ファイル保存用トークン発行
-          $file_token = str_random(32);
-          // 画像ファイル名作成
-          $pictureFile = $file_token . "." . $ext;
-          // 画像ファイル名指定
-          $form['picture'] = $pictureFile;
-          // 画像ファイルをstorage保存
-          $request->picture->storeAs('public/book_images', $pictureFile);
+          // 変更画像情報取得
+          $picture_name = $_FILES['picture']['name']; //ファイル名
+          $picture_ext = substr($picture_name, strrpos($picture_name, '.') + 1); //拡張子
+          $picture_tmp_name = $_FILES['picture']['tmp_name']; //tmp_name
+          $picture_error = $_FILES['picture']['error']; //errorコード
+
+          // 画像ファイルの判定
+          picture_check($picture_ext,$picture_error);
+          // 開発環境で画像保存先を変更
+          if ( app()->isLocal() || app()->runningUnitTests() ) {
+            // ローカル保存処理
+            $request->picture->storeAs('public/'.$save_directory, $picture_name); // 画像ファイルをstorage保存
+            $picture_upload = "/storage/".$save_directory."/".$picture_name; //画像保存パス
+          } else {
+            // 本番環境画像追加処理
+            // S3インスタンス生成
+            $s3settings = s3settings();
+            // S3アップロード処理
+            $picture_upload = picture_upload($save_directory,$picture_name,$picture_tmp_name,$picture_ext,$s3settings);
+          }
         }
+        $form['picture'] = $picture_upload; //画像パスをDB保存値に設定
       }
 
       // レコードアップデート
@@ -161,20 +185,24 @@ class BookController extends Controller
       $save_directory = "book_images";
       // 削除レコード取得
       $delete_book = Bookdata::find($id);
-      // 開発環境で画像保存先を変更
-      if ( app()->isLocal() || app()->runningUnitTests() ) {
-        // 写真削除情報取得
-        $deletename = str_replace('/storage/'.$save_directory.'/','',$delete_book->picture);
 
-        $pathdel = storage_path() . '/app/public/book_images/' . $deletename;
-        // 写真削除
-        \File::delete($pathdel);
-      } else {
-        // 本番環境削除処理
-        // S3インスタンス生成
-        $s3settings = s3settings();
-        // S3削除処理
-        $picture_delete = picture_delete($save_directory,$delete_book,$s3settings);
+      // 写真削除がある場合
+      if (isset($delete_book['picture'])) {
+        // 開発環境で画像保存先を変更
+        if ( app()->isLocal() || app()->runningUnitTests() ) {
+          // 写真削除情報取得
+          $deletename = str_replace('/storage/'.$save_directory.'/','',$delete_book->picture);
+
+          $pathdel = storage_path() . '/app/public/book_images/' . $deletename;
+          // 写真削除
+          \File::delete($pathdel);
+        } else {
+          // 本番環境削除処理
+          // S3インスタンス生成
+          $s3settings = s3settings();
+          // S3削除処理
+          $picture_delete = picture_delete($save_directory,$delete_book,$s3settings);
+        }
       }
       // レコード削除
       $delete_book->delete();
